+++ src/Config/Option/Cors/CorsMaxAgeConfigOption.php
@@ -2,23 +2,24 @@
 
 declare(strict_types=1);
 
-namespace Shlinkio\Shlink\Installer\Config\Option\UrlShortener;
+namespace Shlinkio\Shlink\Installer\Config\Option\Cors;
 
+use Shlinkio\Shlink\Installer\Config\Option\BaseConfigOption;
 use Shlinkio\Shlink\Installer\Config\Util\ConfigOptionsValidator;
 use Symfony\Component\Console\Style\StyleInterface;
 
-class RedirectCacheLifeTimeConfigOption extends AbstractPermanentRedirectDependentConfigOption
+class CorsMaxAgeConfigOption extends BaseConfigOption
 {
     public function getEnvVar(): string
     {
-        return 'REDIRECT_CACHE_LIFETIME';
+        return 'CORS_MAX_AGE';
     }
 
     public function ask(StyleInterface $io, array $currentOptions): int
     {
         return $io->ask(
-            'How long (in seconds) do you want your redirects to be cached by visitors?',
-            '30',
+            'How long (in seconds) do you want CORS config to be cached by browsers?',
+            '3600',
             ConfigOptionsValidator::validatePositiveNumber(...),
         );
     }