+++ src/Config/Option/Database/DatabaseUserConfigOption.php
@@ -4,13 +4,11 @@ declare(strict_types=1);
 
 namespace Shlinkio\Shlink\Installer\Config\Option\Database;
 
-use Shlinkio\Shlink\Installer\Util\AskUtilsTrait;
+use Shlinkio\Shlink\Installer\Config\Util\ConfigOptionsValidator;
 use Symfony\Component\Console\Style\StyleInterface;
 
 class DatabaseUserConfigOption extends AbstractNonSqliteDependentConfigOption
 {
-    use AskUtilsTrait;
-
     public function getEnvVar(): string
     {
         return 'DB_USER';
@@ -18,6 +16,9 @@ class DatabaseUserConfigOption extends AbstractNonSqliteDependentConfigOption
 
     public function ask(StyleInterface $io, array $currentOptions): string
     {
-        return $this->askRequired($io, 'username', 'Database username');
+        return $io->ask(
+            'Database username',
+            validator: static fn ($value) => ConfigOptionsValidator::validateRequired($value, 'username'),
+        );
     }
 }