+++ src/Config/Option/Mercure/MercureJwtSecretConfigOption.php
@@ -4,13 +4,11 @@ declare(strict_types=1);
 
 namespace Shlinkio\Shlink\Installer\Config\Option\Mercure;
 
-use Shlinkio\Shlink\Installer\Util\AskUtilsTrait;
+use Shlinkio\Shlink\Installer\Config\Util\ConfigOptionsValidator;
 use Symfony\Component\Console\Style\StyleInterface;
 
 class MercureJwtSecretConfigOption extends AbstractMercureEnabledConfigOption
 {
-    use AskUtilsTrait;
-
     public function getEnvVar(): string
     {
         return 'MERCURE_JWT_SECRET';
@@ -18,6 +16,9 @@ class MercureJwtSecretConfigOption extends AbstractMercureEnabledConfigOption
 
     public function ask(StyleInterface $io, array $currentOptions): string
     {
-        return $this->askRequired($io, 'JWT secret', 'The secret key known by the mercure hub server to validate JWTs');
+        return $io->ask(
+            'The secret key known by the mercure hub server to validate JWTs',
+            validator: static fn ($value) => ConfigOptionsValidator::validateRequired($value, 'JWT secret'),
+        );
     }
 }