+++ src/Config/Option/RabbitMq/RabbitMqHostConfigOption.php
@@ -4,13 +4,11 @@ declare(strict_types=1);
 
 namespace Shlinkio\Shlink\Installer\Config\Option\RabbitMq;
 
-use Shlinkio\Shlink\Installer\Util\AskUtilsTrait;
+use Shlinkio\Shlink\Installer\Config\Util\ConfigOptionsValidator;
 use Symfony\Component\Console\Style\StyleInterface;
 
 class RabbitMqHostConfigOption extends AbstractRabbitMqEnabledConfigOption
 {
-    use AskUtilsTrait;
-
     public function getEnvVar(): string
     {
         return 'RABBITMQ_HOST';
@@ -18,6 +16,10 @@ class RabbitMqHostConfigOption extends AbstractRabbitMqEnabledConfigOption
 
     public function ask(StyleInterface $io, array $currentOptions): string
     {
-        return $this->askRequired($io, 'RabbitMQ host name');
+        $option = 'RabbitMQ host name';
+        return $io->ask(
+            $option,
+            validator: static fn ($value) => ConfigOptionsValidator::validateRequired($value, $option),
+        );
     }
 }